Trait move_vm_runtime::data_cache::RemoteCache

pub trait RemoteCache {
    fn get_module(&self, module_id: &ModuleId) -> VMResult<Option<Vec<u8>>>;

    fn get_resource(
        &self, 
        address: &AccountAddress, 
        tag: &StructTag
    ) -> PartialVMResult<Option<Vec<u8>>>;
}

Trait for the Move VM to abstract storage operations.

Storage backends should return

Ok(Some(..)) if the data exists

Ok(None) if the data does not exist

Err(..) only when something really wrong happens, for example - invariants are broken and observable from the storage side (this is not currently possible as ModuleId and StructTag are always structurally valid) - storage encounters internal errors

Move VM on the other hand, should NOT blindly trust the storage impl and assume the protocol above is honored. When receiving an error from storage, the Move VM MUST catch the error and convert it into an invariant violation, no matter what type the original error has before propagating the error back to the VM caller.

Eventually we should replace (Partial)VMError with a dedicated VMStorageError or an associated error type so that storage implementations will no longer be able to return a bogus VMError.

Required methods

fn get_module(&self, module_id: &ModuleId) -> VMResult<Option<Vec<u8>>>

fn get_resource(
&self,
address: &AccountAddress,
tag: &StructTag
) -> PartialVMResult<Option<Vec<u8>>>

Loading content...